+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.1] - 26-08-2020
+### Fixed
+- Fixed minor NPE
+
+
+## [4.2.0] - 21-08-2020
+### Added
+- Added collaborative decks (Overloaded)
+- Added cards count for custom decks
+
+### Fixed
+- Fixed some CrCast issues
+- Fixed Overloaded sync issues
+- Other minor fixes
+
+
 ## [4.1.3] - 18-08-2020
 ### Fixed
 - Fixed app crashing on startup (again)